(MODULES-5977) convert acceptance tests to beaker rspec
As per our decision to change all beaker acceptance tests to beaker rspec tests to facilitate consumption of test results by QAELK, this change moves all tests previously in ../acceptance to ../spec/acceptance and converts from straight beaker to beaker rspec.

The invocation remains 'bundle exec rake beaker'
