--- acceptance/.beaker-masterless.cfg
@@ -1,12 +1,0 @@
-{
-    :load_path  => "./acceptance/lib/",
-    :hosts_file => './acceptance/config/windows-2012r2-x86_64a.cfg',
-    :type       => "aio",
-    :helper     => "./acceptance/lib/helper.rb",
-    :pre_suite  => ["./acceptance/setup/00_install_certs.rb","./acceptance/setup/install_puppet.rb","./acceptance/setup/install_modules.rb"],
-    :tests      => "./acceptance/tests",
-    :log_level  => "debug",
-    :timeout    => 6000,
-    :ntp        => true,
-    :keyfile    => "~/.ssh/id_rsa-acceptance"
-}
