--- acceptance/.beaker-pe.cfg
@@ -1,11 +1,0 @@
-{
-    :hosts_file     => './acceptance/config/windows-2012-x86_64.cfg',
-    :type       => "pe",
-    :pre_suite  => ['./acceptance/setup/00_install_certs.rb','./acceptance/setup/install.rb','./acceptance/setup/install_modules.rb'],
-    :tests      => "./acceptance/tests",
-    :helper     => "./acceptance/lib/helper.rb",
-    :log_level  => "debug",
-    :timeout    => 6000,
-    :ntp        => true,
-    :keyfile    => "~/.ssh/id_rsa-acceptance"
-}
