--- acceptance/Gemfile
@@ -1,7 +1,0 @@
-source ENV['GEM_SOURCE'] || "https://rubygems.org"
-
-gem "beaker"
-
-if File.exists? "#{__FILE__}.local"
-  eval(File.read("#{__FILE__}.local"), binding)
-end
