--- acceptance/config/centos-70-x86_64.cfg
@@ -1,12 +1,0 @@
-HOSTS:
-  centos-70-x86_64:
-    roles:
-      - master
-      - database
-      - dashboard
-      - agent
-    platform: el-7-x84_64
-    box: puppetlabs/centos-7.0-64-nocm
-    box_url: https://vagrantcloud.com/puppetlabs/boxes/centos-7.0-64-nocm
-    hypervisor: vagrant
-    pe_dir: http://neptune.puppetlabs.lan/3.4/ci-ready/
