--- acceptance/config/debian-609-amd64.cfg
@@ -1,12 +1,0 @@
-HOSTS:
-  debian-609-amd64:
-    roles:
-      - master
-      - database
-      - dashboard
-      - agent
-    platform: debian-6-amd64
-    box: puppetlabs/debian-6.0.9-64-nocm
-    box_url: https://vagrantcloud.com/puppetlabs/debian-6.0.9-64-nocm
-    hypervisor: vagrant
-    pe_dir: http://neptune.puppetlabs.lan/3.4/ci-ready/
