--- acceptance/config/fedora-18-x86_64.cfg
@@ -1,12 +1,0 @@
-HOSTS:
-  fedora-18-x86_64:
-    roles:
-      - master
-      - database
-      - dashboard
-      - agent
-    platform: fedora-18-x84_64
-    box: fedora-18-x64-vbox4210-nocm
-    box_url: http://puppet-vagrant-boxes.puppetlabs.com/fedora-18-x64-vbox4210-nocm.box
-    hypervisor: vagrant
-    pe_dir: http://neptune.puppetlabs.lan/3.4/ci-ready/
