--- acceptance/config/sles-11-x64.cfg
@@ -1,12 +1,0 @@
-HOSTS:
-  sles-11-x64:
-    roles:
-      - master
-      - database
-      - dashboard
-      - agent
-    platform: sles-11-x64
-    box: sles-11sp1-x64-vbox4210-nocm
-    box_url: http://puppet-vagrant-boxes.puppetlabs.com/sles-11sp1-x64-vbox4210-nocm.box
-    hypervisor: vagrant
-    pe_dir: http://neptune.puppetlabs.lan/3.4/ci-ready/
