--- acceptance/config/super-suite.cfg
@@ -1,47 +1,0 @@
-HOSTS:
-  centos6:
-    roles:
-      - master
-      - database
-      - agent
-      - dashboard
-    platform: el-6-x86_64
-    template: centos-6-x86_64
-    hypervisor: vcloud
-    pe_dir: http://neptune.puppetlabs.lan/3.2/ci-ready/
-  w2k3r2x86:
-    roles:
-      - agent
-    platform: windows-2003r2-i386
-    template: win-2003r2-i386
-    hypervisor: vcloud
-    pe_dir: http://neptune.puppetlabs.lan/3.2/ci-ready/
-  w2k3r2x64:
-    roles:
-      - agent
-    platform: windows-2003r2-x86_64
-    template: win-2003r2-x86_64
-    hypervisor: vcloud
-    pe_dir: http://neptune.puppetlabs.lan/3.2/ci-ready/
-  w2k8r2:
-    roles:
-      - agent
-    platform: windows-2008r2-x86_64
-    template: win-2008r2-x86_64
-    hypervisor: vcloud
-    pe_dir: http://neptune.puppetlabs.lan/3.2/ci-ready/
-  w2012:
-    roles:
-      - agent
-      - default
-    platform: windows-2012-x86_64
-    template: win-2012-x86_64
-    hypervisor: vcloud
-    pe_dir: http://neptune.puppetlabs.lan/3.2/ci-ready/
-CONFIG:
-  nfs_server: none
-  consoleport: 443
-  datastore: instance0
-  folder: Delivery/Quality Assurance/FOSS/Dynamic
-  resourcepool: delivery/Quality Assurance/FOSS/Dynamic
-  pooling_api: http://vcloud.delivery.puppetlabs.net/
