--- acceptance/config/ubuntu-1204-amd64.cfg
@@ -1,12 +1,0 @@
-HOSTS:
-  ubuntu-1204-amd64:
-    roles:
-      - master
-      - database
-      - dashboard
-      - agent
-    platform: ubuntu-12.04-amd64
-    box: puppetlabs/ubuntu-12.04-64-nocm
-    box_url: https://vagrantcloud.com/puppetlabs/ubuntu-12.04-64-nocm
-    hypervisor: vagrant
-    pe_dir: http://neptune.puppetlabs.lan/3.4/ci-ready/
