--- acceptance/config/windows-2003r2-x86_64.cfg
@@ -1,24 +1,0 @@
-HOSTS:
-  debian6:
-    roles:
-      - master
-      - database
-      - agent
-      - dashboard
-    platform: debian-6-i386
-    template: debian-6-i386
-    hypervisor: vcloud
-  w2k3r2:
-    roles:
-      - agent
-    platform: windows-2003r2-x86_64
-    template: win-2003r2-x86_64
-    hypervisor: vcloud
-CONFIG:
-  pe_dir: http://neptune.puppetlabs.lan/3.2/ci-ready/
-  nfs_server: none
-  consoleport: 443
-  datastore: instance0
-  folder: Delivery/Quality Assurance/FOSS/Dynamic
-  resourcepool: delivery/Quality Assurance/FOSS/Dynamic
-  pooling_api: http://vcloud.delivery.puppetlabs.net/
