--- acceptance/config/windows-2012-x86_64.cfg
@@ -1,26 +1,0 @@
-HOSTS:
-  debian6:
-    roles:
-      - master
-      - database
-      - agent
-      - dashboard
-    platform: debian-6-i386
-    template: debian-6-i386
-    hypervisor: vcloud
-  w2012:
-    roles:
-      - agent
-      - default
-    platform: windows-2012-x86_64
-    template: win-2012-x86_64
-    hypervisor: vcloud
-CONFIG:
-  pe_dir: http://neptune.puppetlabs.lan/3.4/ci-ready/
-  nfs_server: none
-  consoleport: 443
-  datastore: instance0
-  folder: Delivery/Quality Assurance/FOSS/Dynamic
-  resourcepool: delivery/Quality Assurance/FOSS/Dynamic
-  pooling_api: http://vcloud.delivery.puppetlabs.net/
-  keyfile: ~/.ssh/id_rsa-acceptance
