--- acceptance/config/windows-2012r2-x86_64a.cfg
@@ -1,16 +1,0 @@
-HOSTS:
-  win-2012r2-x86_64-agent:
-    roles:
-      - agent
-      - default
-    platform: windows-2012r2-x86_64
-    template: Delivery/Quality Assurance/Templates/vCloud/win-2012r2-wmf5-x86_64
-    hypervisor: vcloud
-CONFIG:
-  masterless: true
-  nfs_server: none
-  consoleport: 443
-  datastore: instance0
-  folder: Delivery/Quality Assurance/Enterprise/Dynamic
-  resourcepool: delivery/Quality Assurance/Enterprise/Dynamic
-  pooling_api: http://vmpooler.delivery.puppetlabs.net/
