--- acceptance/setup/install.rb
@@ -1,3 +1,0 @@
-test_name "Installing Puppet Enterprise" do
-  install_pe
-end
