--- acceptance/setup/install_puppet.rb
@@ -1,5 +1,0 @@
-require 'beaker/puppet_install_helper'
-
-test_name "Installing Puppet" do
-  run_puppet_install_helper
-end
