+++ spec/acceptance/nodesets/hosts.yml
@@ -1,0 +1,31 @@
+HOSTS:
+  windows2008r2-64-1:
+    pe_dir:
+    pe_ver:
+    pe_upgrade_dir:
+    pe_upgrade_ver:
+    hypervisor: abs
+    platform: windows-2008r2-64
+    packaging_platform: windows-2012-x64
+    ruby_arch: x64
+    template: win-2008r2-x86_64
+    roles:
+    - agent
+    - default
+  redhat7-64-1:
+    pe_dir:
+    pe_ver:
+    pe_upgrade_dir:
+    pe_upgrade_ver:
+    hypervisor: abs
+    platform: el-7-x86_64
+    packaging_platform: el-7-x86_64
+    template: redhat-7-x86_64
+    roles:
+    - agent
+    - master
+    - database
+    - dashboard
+CONFIG:
+  nfs_server: none
+  consoleport: 443
